% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftb.R
\name{int2word}
\alias{int2word}
\title{Converts integers into English words,
Based on a function by John Fox:
http://tolstoy.newcastle.edu.au/R/help/05/04/2715.html}
\usage{
int2word(x)
}
\arguments{
\item{x}{an integer}
}
\description{
Converts integers into English words,
Based on a function by John Fox:
http://tolstoy.newcastle.edu.au/R/help/05/04/2715.html
}
\examples{
in2word(324513)
}
