% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_spacer.R
\name{int_spacer}
\alias{int_spacer}
\title{Round a number to the nearest integer and
convert to a string with a non-breaking space
as the thousands separator. Do not use
the thousands separator for numbers less than ten thousand}
\usage{
int_spacer(x)
}
\arguments{
\item{x}{A number to be rounded and spaced}
}
\value{
A string
}
\description{
Also allows for the input to be a string containing a numeric
}
\examples{
int_spacer(2.4)
int_spacer(24.5)
int_spacer(9876)
int_spacer(1234567.8)
int_spacer("123456789")

}
