% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_gtb.R
\name{palette_gtb}
\alias{palette_gtb}
\title{Colour palettes for variables that appear in multiple charts in the
WHO Global TB Report}
\usage{
palette_gtb(var_name)
}
\arguments{
\item{var_name}{Name of the variable being plotted}
}
\value{
A hexadecimal colour code, or an empty string if var_name was not recognised.
}
\description{
Colour palettes for variables that appear in multiple charts in the
WHO Global TB Report
}
\note{
var_name is not case sensitve. It must be one of:

TB burden estimates:

- "inc"    (incidence)

- "inch"   (HIV-positive incidence)

- "mort"   (mortality)



Disaggregation by sex:

- "female"

- "male"



Disaggregated by WHO region:

- "afr"

- "amr"

- "emr"

- "eur"

- "sear"

- "wpr"



Treatment outcomes:

- "succ"   (treatment success)

- "fail"   (treatment failure)

- "died"   (died)

- "ltfu"   (lost to follow-up)

- "neval"  (not evaluated)

- "no_outcomes" (no outcomes data reported)
}
\examples{
p <- ggplot2::ggplot(mtcars, aes(wt, mpg)) +

              geom_point(size = 3, colour = palette_gtb("inc")) +

              theme_gtb()
print(p)

}
\author{
: Hazim Timimi
}
