% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gtb.R
\name{theme_gtb}
\alias{theme_gtb}
\title{ggplot2 theme for the online WHO Global TB Report}
\usage{
theme_gtb(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
ggplot2 theme for the online WHO Global TB Report
}
\author{
: Hazim Timimi
}
