% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftb.R
\name{ftb}
\alias{ftb}
\title{Number formatter according to GTB rounding rules}
\usage{
ftb(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
Formats vectors of numbers (<2 billion)
}
\examples{
ftb(348838)
ftb(0.0359)
ftb(0.00036)

}
