% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_spacer.R
\name{int_spacer}
\alias{int_spacer}
\title{Round a number to the nearest integer and
convert to a string with space as the
thousands separator}
\usage{
int_spacer(x)
}
\arguments{
\item{x}{A number to be rounded and spaced}
}
\value{
A string
}
\description{
Round a number to the nearest integer and
convert to a string with space as the
thousands separator
}
\examples{
int_spacer(23.4)
int_spacer(9876)
int_spacer(1234567.8)

}
