% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_spacer.R
\name{int_spacer}
\alias{int_spacer}
\title{Round a number to the nearest integer and format according to WHO style guide}
\usage{
int_spacer(x)
}
\arguments{
\item{x}{A number to be rounded and spaced}
}
\value{
A string
}
\description{
Round a number to the nearest integer and format according to WHO style guide.
This means using non-breaking spaces as thousands separators, but don;t do that
for numbers between 1000 and 9999
}
\examples{
int_spacer(2.4)
int_spacer(24.5)
int_spacer(9876)
int_spacer(1234567.8)
int_spacer(c(123456, 7890123))

}
